services/service-b/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-asgi
prometheus-fastapi-instrumentator